+++ torchmetrics/__init__.py
@@ -57,11 +57,13 @@ from torchmetrics.regression import (  # noqa: E402
 )
 from torchmetrics.retrieval import (  # noqa: E402
     RetrievalFallOut,
+    RetrievalHitRate,
     RetrievalMAP,
     RetrievalMRR,
     RetrievalNormalizedDCG,
     RetrievalPrecision,
     RetrievalRecall,
+    RetrievalRPrecision,
 )
 from torchmetrics.text import WER, BERTScore, BLEUScore, ROUGEScore, SacreBLEUScore  # noqa: E402
 from torchmetrics.wrappers import BootStrapper, MetricTracker, MinMaxMetric, MultioutputWrapper  # noqa: E402
@@ -117,11 +119,13 @@ __all__ = [
     "R2Score",
     "Recall",
     "RetrievalFallOut",
+    "RetrievalHitRate",
     "RetrievalMAP",
     "RetrievalMRR",
     "RetrievalNormalizedDCG",
     "RetrievalPrecision",
     "RetrievalRecall",
+    "RetrievalRPrecision",
     "ROC",
     "ROUGEScore",
     "SacreBLEUScore",