+++ torchmetrics/functional/__init__.py
@@ -58,8 +58,10 @@ from torchmetrics.functional.regression.symmetric_mean_absolute_percentage_error
 from torchmetrics.functional.regression.tweedie_deviance import tweedie_deviance_score
 from torchmetrics.functional.retrieval.average_precision import retrieval_average_precision
 from torchmetrics.functional.retrieval.fall_out import retrieval_fall_out
+from torchmetrics.functional.retrieval.hit_rate import retrieval_hit_rate
 from torchmetrics.functional.retrieval.ndcg import retrieval_normalized_dcg
 from torchmetrics.functional.retrieval.precision import retrieval_precision
+from torchmetrics.functional.retrieval.r_precision import retrieval_r_precision
 from torchmetrics.functional.retrieval.recall import retrieval_recall
 from torchmetrics.functional.retrieval.reciprocal_rank import retrieval_reciprocal_rank
 from torchmetrics.functional.self_supervised import embedding_similarity
@@ -114,8 +116,10 @@ __all__ = [
     "recall",
     "retrieval_average_precision",
     "retrieval_fall_out",
+    "retrieval_hit_rate",
     "retrieval_normalized_dcg",
     "retrieval_precision",
+    "retrieval_r_precision",
     "retrieval_recall",
     "retrieval_reciprocal_rank",
     "roc",